✨ Refactor application review stats handling
- Added `ApplicationReviewStats` record for application review metrics.
- Updated `GetApplicationReviewsCountApiResponse` to use the new record.
- Modified implicit conversion in `DashboardApplicationReviewStats`.
- Changed return types in `IEmployerAccountProvider` and `ITrainingProviderService`.
- Updated tests to reflect new return types and cleaned up unused usings.

Changes made by Balaji Jambulingam
