+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Services/EmployerAccount/IEmployerAccountProvider.cs
@@ -2,7 +2,6 @@
 using System.Threading.Tasks;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.OuterApi.Responses;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.QueryStore.Projections.EditVacancyInfo;
-using Esfa.Recruit.Vacancies.Client.Infrastructure.ReferenceData.Dashboard;
 
 namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services.EmployerAccount
 {
@@ -12,7 +11,7 @@ namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services.EmployerAccount
         Task<IEnumerable<LegalEntity>> GetEmployerLegalEntitiesAsync(string hashedAccountId);
         Task<string> GetEmployerAccountPublicHashedIdAsync(string hashedAccountId);
         Task<IEnumerable<AccountLegalEntity>> GetLegalEntitiesConnectedToAccountAsync(string hashedAccountId);
-        Task<DashboardApplicationReviewStats> GetEmployerDashboardApplicationReviewStats(string hashedAccountId,
+        Task<List<ApplicationReviewStats>> GetEmployerDashboardApplicationReviewStats(string hashedAccountId,
             List<long> vacancyReferences);
     }
 }
